Move codec utils into separate codec package
