+++ pkg/codec/map_decoder.go
@@ -1,4 +1,4 @@
-package utils
+package codec
 
 import (
 	"context"
